graphql: move the users queries to their own module
